# api/swagger/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"DefaultSelection_Prisma._36_UserPayload_": {
				"properties": {
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"phone": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"hashPwd": {
						"type": "string"
					},
					"fullname": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"updatedAt",
					"createdAt",
					"phone",
					"email",
					"hashPwd",
					"fullname",
					"username",
					"id"
				],
				"type": "object"
			},
			"PrismaUser": {
				"$ref": "#/components/schemas/DefaultSelection_Prisma._36_UserPayload_",
				"description": "Model User"
			},
			"DefaultSelection_Prisma._36_Power655DrawPayload_": {
				"properties": {
					"jackpot2": {
						"type": "string"
					},
					"jackpot1": {
						"type": "string"
					},
					"wns": {
						"type": "string"
					},
					"drawId": {
						"type": "number",
						"format": "double"
					},
					"date": {
						"type": "string",
						"format": "date-time"
					},
					"t": {
						"type": "string"
					},
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"jackpot2",
					"jackpot1",
					"wns",
					"drawId",
					"date",
					"t",
					"id"
				],
				"type": "object"
			},
			"Power655Draw": {
				"$ref": "#/components/schemas/DefaultSelection_Prisma._36_Power655DrawPayload_",
				"description": "Model Power655Draw"
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "api",
		"version": "1.0.0",
		"description": "A description of POMELO API",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/users": {
			"get": {
				"operationId": "GetUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/PrismaUser"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"201": {
						"description": "User created",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"user": {
											"$ref": "#/components/schemas/PrismaUser"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"user",
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"phone": {
										"type": "string"
									},
									"email": {
										"type": "string"
									},
									"hashPwd": {
										"type": "string"
									},
									"fullname": {
										"type": "string"
									},
									"username": {
										"type": "string"
									}
								},
								"required": [
									"phone",
									"email",
									"hashPwd",
									"fullname",
									"username"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/users/{id}": {
			"get": {
				"operationId": "GetUserById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/PrismaUser"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"404": {
						"description": "User not found",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"enum": [
										null
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/": {
			"get": {
				"operationId": "WelcomeMessage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/power655": {
			"get": {
				"operationId": "WelcomeMessage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Power655"
				],
				"security": [],
				"parameters": []
			}
		},
		"/power655/results": {
			"get": {
				"operationId": "GetResultAll",
				"responses": {
					"200": {
						"description": "Get latest result from database",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Power655Draw"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Power655"
				],
				"security": [],
				"parameters": []
			},
			"delete": {
				"operationId": "DropResultAll",
				"responses": {
					"200": {
						"description": "All results deleted successfully"
					}
				},
				"description": "Deletes all records from the Power655Draw table.\nThis operation will permanently remove all draw results from the database.",
				"summary": "Delete all Power655 draw results from the database",
				"tags": [
					"Power655"
				],
				"security": [],
				"parameters": []
			}
		},
		"/power655/results/last/{quantity}": {
			"get": {
				"operationId": "GetResultByLastDrawIds",
				"responses": {
					"200": {
						"description": "Get latest result from database",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Power655Draw"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Power655"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "quantity",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/power655/results/range": {
			"get": {
				"operationId": "GetResultByDateRange",
				"responses": {
					"200": {
						"description": "Get results by date range 2",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Power655Draw"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Power655"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "startDate",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "endDate",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/power655/results/{lastDrawId}": {
			"get": {
				"operationId": "GetResultFrom3rdParty",
				"responses": {
					"200": {
						"description": "Get latest result from 3rd party",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Power655Draw"
									},
									"type": "array",
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Power655"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "lastDrawId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/power655/results/sync/{lastDrawId}": {
			"post": {
				"operationId": "SyncResultAllFrom3rdParty",
				"responses": {
					"201": {
						"description": "Sync latest result from 3rd party",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Power655Draw"
									},
									"type": "array",
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Power655"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "lastDrawId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/api"
		}
	]
}